+++ src/main/kotlin/com/example/myapp/repository/UserRepository.kt
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface UserRepository : JpaRepository<User, Long> {
+interface UserRepository : JpaRepository<User, Int> {
     fun findByUsername(username: String): User?
     fun findByEmail(email: String): User?
     fun existsByUsername(username: String): Boolean
